+++ packages/plugin-ext/src/hosted/node/hosted-plugin.ts
@@ -16,7 +16,7 @@
 
 import { injectable, inject, multiInject, postConstruct, optional } from '@theia/core/shared/inversify';
 import { ILogger, ConnectionErrorHandler } from '@theia/core/lib/common';
-import { HostedPluginClient, PluginModel, ServerPluginRunner, DeployedPlugin, PluginIdentifiers } from '../../common/plugin-protocol';
+import { HostedPluginClient, PluginModel, ServerPluginRunner } from '../../common/plugin-protocol';
 import { LogPart } from '../../common/types';
 import { HostedPluginProcess } from './hosted-plugin-process';
 
@@ -90,20 +90,6 @@ export class HostedPluginSupport {
             this.hostedPluginProcess.runPluginServer(serverName);
             this.isPluginProcessRunning = true;
         }
-    }
-
-    /**
-     * Provides additional plugin ids.
-     */
-    async getExtraDeployedPluginIds(): Promise<PluginIdentifiers.VersionedId[]> {
-        return [].concat.apply([], await Promise.all(this.pluginRunners.map(runner => runner.getExtraDeployedPluginIds())));
-    }
-
-    /**
-     * Provides additional deployed plugins.
-     */
-    async getExtraDeployedPlugins(): Promise<DeployedPlugin[]> {
-        return [].concat.apply([], await Promise.all(this.pluginRunners.map(runner => runner.getExtraDeployedPlugins())));
     }
 
     sendLog(logPart: LogPart): void {